+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
     Created Package
     [x] Work on Documentation
-    [x] Provide more flexibility arounf Filter View
+    [x] Provide more flexibility around Filter View
+
+# 2.0.0
+
+    Added more flexibility around Filter Component